// MongoBooks2/JavaScript/Node/AngularCLI/my-project/src/app/Models/BookData.ts
export class BookData {
  "_id": string;
  "dateString": string;
  "date": Date;
  "author": string;
  "title": string;
  "pages": number;
  "note": string;
  "nationality": string;
  "originalLanguage": string;
  "image_url": string;
  "tags": Array<string>;
  "Book": string;
  "Comic": string;
  "Audio": string;
  "format": number;
  "user": any;  
}